requirements.txt
pandas
pyarrow
scikit-learn
transformers
datasets
torch
shap
mlflow
fastapi
uvicorn[standard]
pydantic
redis
slowapi
python-jose[cryptography]
streamlit
plotly
evidently
fpdf2
joblib
pytest
httpx
prometheus-fastapi-instrumentator
python-dotenv
psycopg2-binary
pyyaml
sentencepiece
tiktoken
openai>=1.0.0
seaborn
matplotlib
accelerate
